update fiscal EfdRegistros

+++ backend/src/main/java/com/example/backend/fiscal/efdRegistros/EfdRegistrosRepository.java
@@ -2,5 +2,5 @@ package com.example.backend.fiscal.efdRegistros;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EfdRegistrosRepository extends JpaRepository<EfdRegistros, Integer> {
+public interface EfdRegistrosRepository extends JpaRepository<EfdRegistros, Long> {
 }
